survey export fixes, ui tidy
