+++ Decsys/ClientApp/src/app/components/ExportModal/index.js
@@ -1,0 +1,2 @@
+import ExportModal from "./ExportModal";
+export default ExportModal;
